increments instead of bigincrements

+++ database/migrations/2020_02_09_201925_create_app_answers_table.php
@@ -15,8 +15,8 @@ class CreateAppAnswersTable extends Migration
     {
         Schema::create('app_answers', function (Blueprint $table) {
             $table->bigIncrements('id');
-            $table->bigInteger('app_id')->unsigned();
-            $table->bigInteger('question_id')->unsigned();
+            $table->integer('app_id')->unsigned();
+            $table->integer('question_id')->unsigned();
             $table->string('answer', 3000);
             $table->timestamps();
